Refactor VideoProcessingThread into separate module
